Adding IServiceBusHub and IServiceBusHubFactory.
An IServiceBusHub is an IServiceBus and all the necessary supporting
services to support it. I couldn't think of a better name.

This also means that the ISubscriptionService is no longer singleton
in the applicationi or living in the container. Each hub has it's own
because we can have server hubs and subscription manager hubs.