Converting to a single git repository.
This adds .gitignore but it's the first real commit to the new repository
that I just made. I stripped all the history for the extensions'
subdirectory and placed all of it here. Now I'm adding the necessary
libraries so that I can build.

It's awkward dealing with so many different projects in one git repository
and so I decided to take the plunge and break this out. I'd like to see
that happen for two other things in Machine proper but we'll see. I think
the only real difficulty I'll experience is the added dependency
juggling issues. Thankfully, I'll get Partstore up and running soon...