Separating the older wrapper code from the newer MTA stuff.
I have a lot of older code in here that wraps a lot of the finer grained
setup code for MT to use MT in existing applications and situations that
I'm dealing with. I'm keeping this wrapper code around for the time being.

It's just code to integrate with Machine.Container and allow for easier
hosting of multiple buses in an application. I did this so I didn't have
to run a separate subscription manager, etc... And because I had grand
visions for an application container type framework. I may return to this
later but it'll probably involve the newer MTA stuff anyway.