Refactoring how handlers are discovered and found.
I'm going to add a nifty ordering mechanism and need this to be more
flexible. I've made this so that a HandlerDiscoverer will return a
list of FutureHandlerInvocation's that are then used to perform the
invocations themselves.

I can then change the order of that returned list and tweak that as I
see fit...