Refactoring how the non-generic invokers are created.
Got rid of the separate interface, now we just create a very vauge
implementation that we can pass the message into. For example our wrapper
is a Consumes<IMessage>.All and takes a more specific one. We can then
create one of those using MakeGenericType magic and pass in the version
we'd like them to call. This is much prettier.