Ordering of message handlers.
This looks for an IProvideHandlerOrderFor for the given message type
and then attempts to re-order the future invocations based on the
types that it returns. For now this works but later on we may need to
make this trickier, like supporting the ordering of implementations that
consume the same message on the same target for example.