Renaming my hacks at fixing weirdness in C#'s type system.
These tested more for contravariance then covariance I think. And I made
the naming fuzzier since I'm not sure of the correctness totally. They're
mostly for supporting quirky generics functionality.

This also fixes a bug in that any interface that's contravariant with the
given type also means we're contravariant. This was breaking the case
where something implemented Of<Base> and we were seeing if it was
contravariant with Of<Derrived>. Basically we just check the interfaces
of both sides of the comparison now.