A fully functional timing backend.
This is basically a refactoring and re-implementation of the existing
background task system that we had but made much cleaner. It's not really
a spike since I have a lot of specifications in there.

In an effort to simplfy the only operation we really support is scheduling
via a cron like system when a certain message is published. So there's
some trigger infrastructure (with a CronTrigger) and then a task that
is fired when that trigger is true, in this case one that publishes a
message.

We need the factory because that task needs the bus and the message
factory and all that stuff.