A very simple Saga implementation.
Our Sagas operate using the MessageAspect infrastructure. We check to see
if a handler is a SagaHandler and that turns things on. If we have a
SagaMessage then we look for the saga's state using a repository and
give that to the handler. When we're done invocation the consumer method
we look for state and then save that or delete depending on if the saga
is complete.

Notice this decouples state from the handler (providing a State property
on the handler itself)

I'm not sure if we like having IsSagaComplete on ISagaState or if that
should be on the handler. I'm leaning more towards this for now. We
let the repository decide what to do when we can't find the state for
the saga.