AdoNet repositories/storage for future publishes and sagas.
This is something I spiked last night that let's us store future publishes
and saga state in a SQL database. We use raw/low level SQL. I didn't see
a huge need to deal with NH at all for this type of largely static thing.
For people that need NH saga's they'll have their own infrastructure in
place to do NH stuff. This is just a handy thing that can be dropped in
whenever something reliable/robust is necessary and we don't want to invest
the time in making an NH implementation.