Letting future publishes specify where to send the message.
Instead of just replying back we apply the rules to the scheduled message
that are on the scheduler's side and provide those to the timing
service, which will then send the message to those addresses. It may be
argued that it makes sense for the timing service to apply the rules but
the problem there is that we lose the transparency of the timing service
because right now it's just operating on a raw payload.