Upgrading to the latest MT.
They changed their namespaces, that's a bulk of this commit. Also they
did some file renames and it's easier to just nuke and add with git.

Also had to fix a problem in the contravariance checker for finding
handlers because having common base interfaces was incorrectly causing that
to return true.