Allowing more flexibility in how we specify CorrelationId's.
I'm going to experiment with an easier way to correlate messages in a
saga and it's going to involve overloading CorrelationId on outgoing and
incoming messages.

This also introduces a ReturnCorrelationId, so we can tell the caller
how we'd like to be correlated by. To reproduce the old behavior we
can use the Id.