Adding a more pervasive place to store SagaId.
I just hated that I had to build saga stuff into existing messages when
the problem was just a simple correlation issue. I decided that it was
worthwhile to build this into the TransportMessage itself and interpret
that in addition to any potential ISagaMessage's. It's kind of leaky I
agree, especially with how awesomely separate saga's are now. But, again
this is such an important pattern it's worth the trade off.

The only question then is how do we set this Id to kick things off?