Adding our own IEndpoint and friends, no longer referencing MT.
All we were using of MT was the IEndpoint code and even then I had some
custom changes I had made. My whole reasoning for this project was the
drastic difference between my idea/implementation of a message bus and MT
and now the dependency on MT is starting to make less and less sense. The
amount of code is pretty trivial and the dependence doesn't really give us
much as far as re-use.