Fancier logging. This may be a performance problem eventually.
In the future we may decide to move this into Debug or something and check
to ensure things are enabled before even trying to log stuff. For now
I'm going to leave things be.

This allows us to filter per-message so we can ignore common messages that
just tend to clutter things up.