Allow message creation from anonymous objects.
We can do this now:

factory.Create<ISomeMessage>(new { Id = Guid.NewGuid, Name = "Jacob" });

And the factory will perform validation to ensure the given properties are
on the message and all the properties are provided for correctness.

We also remain backwards compatible as well. All this does is introduce
a new constructor to the generated messages.