CurrentMessageBus thread static that can be used to remove guesswork.
I wanted to be able to easily allow multiple listening buses in a single
process and all the solutions usually required a kind of hub or provider
service to hand the bus to the consuming system. I've decided to
investigate some kind of context based scheme that just makes the right bus
available through some kind of convention? Eh, we'll see.