Got rid of the old ThreadPool integration in favor of a simpler approach.
We now ask the endpoint if its got anything for read so that we can do
that step outside of a transactional scope and then we start a transaction.

I've also kept an eye on being able to get exceptions that occur so that
we can forward a message describing the problem in the actual
implementations and such.